refactor(infra): centralize Lambda database capacity policy
